Xplor: behavior of channels after writers dropped
My expectation matches behavior: that the channel contents are
persisted.
However, this differs from my reading of the tokio docs, which suggests
that the channel should return `None` if there are no valid senders.

I am uncertain if they actually mean `None` if no valid senders AND
channel contents are exhausted, OR if I'm not understanding when the
senders (send handles) are actually dropped.