Fix: inactivated example code; type issue
I forget some of the context of writing this.  Code was never hooked up
to run and had some dependency on external types.  That said, the code
as written takes a BufWriter wrapping a TCPStream somwhere that it says
it would like only the inner value.
Doesn't seem worth it to track down the origin of this.