Feat: dual approaches to actor & channel gen
the `new()` is more complicated than would be
appropriate normally, but useful here for
exploration.

If `new(Some(rx))` only an actor is generated,
with a tx implicitely elsewhere in the wings.

If `new(None)` then actor & actorhandle are gen'd
with tx & rx embedded within.

Also added the task spawning function and small
running encapsulation functions from
`actor_impl_expanded.rs` and then deleted that
file as it was no longer needed